--- packages/ckeditor5-powered-by/package.json
@@ -1,49 +1,0 @@
-{
-  "name": "@ckeditor/ckeditor5-powered-by",
-  "version": "37.1.0",
-  "description": "Powered by feature for CKEditor 5.",
-  "keywords": [
-    "ckeditor",
-    "ckeditor5",
-    "ckeditor 5",
-    "ckeditor5-feature",
-    "ckeditor5-plugin",
-    "ckeditor5-dll"
-  ],
-  "main": "src/index.ts",
-  "dependencies": {
-    "ckeditor5": "^37.1.0"
-  },
-  "devDependencies": {
-    "@ckeditor/ckeditor5-core": "^37.1.0",
-    "@ckeditor/ckeditor5-editor-classic": "^37.1.0",
-    "typescript": "^4.8.4",
-    "webpack": "^5.58.1",
-    "webpack-cli": "^4.9.0"
-  },
-  "engines": {
-    "node": ">=16.0.0",
-    "npm": ">=5.7.1"
-  },
-  "author": "CKSource (http://cksource.com/)",
-  "license": "GPL-2.0-or-later",
-  "homepage": "https://ckeditor.com/ckeditor-5",
-  "bugs": "https://github.com/ckeditor/ckeditor5/issues",
-  "repository": {
-    "type": "git",
-    "url": "https://github.com/ckeditor/ckeditor5.git",
-    "directory": "packages/ckeditor5-powered-by"
-  },
-  "files": [
-    "lang",
-    "src/**/*.js",
-    "src/**/*.d.ts",
-    "theme",
-    "ckeditor5-metadata.json",
-    "CHANGELOG.md"
-  ],
-  "scripts": {
-    "build": "tsc -p ./tsconfig.json",
-    "postversion": "npm run build"
-  }
-}
