--- packages/ckeditor5-powered-by/src/index.ts
@@ -1,12 +1,0 @@
-/**
- * @license Copyright (c) 2003-2023, CKSource Holding sp. z o.o. All rights reserved.
- * For licensing, see LICENSE.md or https://ckeditor.com/legal/ckeditor-oss-license
- */
-
-/**
- * @module undo
- */
-
-export { default as PoweredBy } from './poweredby';
-
-import './augmentation';
