--- packages/ckeditor5-powered-by/tsconfig.json
@@ -1,10 +1,0 @@
-{
-	"extends": "../../tsconfig.release.json",
-	"include": [
-		"src",
-		"../../typings"
-	],
-	"exclude": [
-		"tests"
-	]
-}
