+++ packages/ckeditor5-typing/src/typing.ts
@@ -11,9 +11,6 @@ import { Plugin } from '@ckeditor/ckeditor5-core';
 import Input from './input';
 import Delete from './delete';
 
-// TODO: The way of injecting this plugin into all editors. EditorUI? EditableUIView?
-import PoweredBy from '@ckeditor/ckeditor5-powered-by/src/poweredby';
-
 /**
  * The typing feature. It handles typing.
  *
@@ -22,7 +19,7 @@ import PoweredBy from '@ckeditor/ckeditor5-powered-by/src/poweredby';
  */
 export default class Typing extends Plugin {
 	public static get requires() {
-		return [ Input, Delete, PoweredBy ] as const;
+		return [ Input, Delete ] as const;
 	}
 
 	/**
