+++ packages/ckeditor5-ui/tests/manual/poweredby/poweredby.html
@@ -96,3 +96,36 @@
 <div id="padding-less">
 	<p>Like all the great things on earth traveling teaches us by example. Here are some of the most precious lessons I’ve</p>
 </div>
+
+<h2>Parent with overflow</h2>
+<div style="height: 300px; overflow-y: scroll; padding: 30px; border: 1px solid red;">
+	<div id="overflow-parent">
+		<h2>The three greatest things you learn from traveling</h2>
+		<p>Like all the great things on earth traveling teaches us by example. Here are some of the most precious lessons I’ve
+			learned over the years of traveling.</p>
+
+		<blockquote>
+			<p>The real voyage of discovery consists not in seeking new landscapes, but having new eyes.</p>
+			<p><strong>Marcel Proust</strong></p>
+		</blockquote>
+		<h3>Improvisation</h3>
+		<p>Life doesn't allow us to execute every single plan perfectly. This especially seems to be the case when you travel.
+			You plan it down to every minute with a big checklist. But when it comes to executing it, something always comes up
+			and you’re left with your improvising skills. You learn to adapt as you go. Here’s how my travel checklist looks
+			now:</p>
+		<ul>
+			<li>buy the ticket</li>
+			<li>start your adventure</li>
+		</ul>
+		<figure class="image image-style-side"><img src="https://ckeditor.com/docs/ckeditor5/latest/assets/img/umbrellas.jpg"
+				alt="Three monks ascending the stairs of an ancient temple."
+				srcset="https://ckeditor.com/docs/ckeditor5/latest/assets/img/umbrellas.jpg, https://ckeditor.com/docs/ckeditor5/latest/assets/img/umbrellas_2x.jpg 2x"
+				sizes="100vw">
+			<figcaption>Three monks ascending the stairs of an ancient temple.</figcaption>
+		</figure>
+		<h3>Confidence</h3>
+		<p>Going to a new place can be quite terrifying. While change and uncertainty make us scared, traveling teaches us how
+			ridiculous it is to be afraid of something before it happens. The moment you face your fear and see there is nothing
+			to be afraid of, is the moment you discover bliss.</p>
+	</div>
+</div>
