+++ packages/ckeditor5-ui/tests/manual/poweredby/poweredby.md
@@ -1,0 +1,6 @@
+# Link to project's homepage rendered next to the editing root.
+
+1. Make sure the "powered by" link renders next to the bottom edge of the editing root when focused.
+1. Make sure the link opens the page in a new tab.
+1. Make sure the link disappears as soon as the editor gets blurred.
+1. Make sure the link disappears when the editor is cropped by an ancestor with `overflow`.
