+++ packages/ckeditor5-ui/theme/globals/globals.css
@@ -7,3 +7,4 @@
 @import "./_reset.css";
 @import "./_zindex.css";
 @import "./_transition.css";
+@import "./_poweredby.css";
